+++ src/main/java/com/nttdata/accounts/service/PagoMensual.java
@@ -3,6 +3,6 @@ package com.nttdata.accounts.service;
 @FunctionalInterface
 public interface PagoMensual {
 
-    double calcular(float amount, float tea, float time);
+    double calcular(Double amount, Double tea, Double time);
 
 }
