+++ src/main/java/com/nttdata/accounts/service/impl/CreditServiceImpl.java
@@ -31,7 +31,7 @@ public class CreditServiceImpl implements CreditService {
         System.out.printf("Mensual sin interes: " + entity.getAmount()/entity.getFeeMonthly());
         Double pagoMensual = op.calcular(entity.getAmount(), entity.getTea(), entity.getFeeMonthly());
         System.out.printf("Mensual con interes: " + pagoMensual);
-        entity.setFeeMonthly(pagoMensual.floatValue());
+        entity.setFeeMonthly(pagoMensual);
         entity.setStatus(true);
         return creditRepository.save(entity);
     }
@@ -61,7 +61,7 @@ public class CreditServiceImpl implements CreditService {
                     origin.setAmount(entity.getAmount());
                     origin.setTea(entity.getTea());
                     origin.setFeeMonthly(entity.getFeeMonthly());
-                    origin.setFeeMonthly(pagoMensual.floatValue());
+                    origin.setFeeMonthly(pagoMensual);
                     origin.setStatus(entity.getStatus());
                     return creditRepository.save(origin);
                 });
